fix: refactors refresh token entity
Moves refresh token entity and repository to a core package.
Removes unused fields and adds builder pattern for simplified creation.
