+++ src/main/java/com/lp/v2/domains/auth/core/RefreshTokenEntity.java
@@ -1,8 +1,9 @@
-package com.lp.v2.domains.auth;
+package com.lp.v2.domains.auth.core;
 
 import com.lp.v2.common.Auditable;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -24,10 +25,9 @@ public class RefreshTokenEntity extends Auditable {
     @Column(name = "account_id", nullable = false)
     private Long accountId;
 
-    @Column(name = "issued_at", nullable = false)
-    private LocalDateTime issuedAt;
-
-    @Column(name = "expires_at", nullable = false)
-    private LocalDateTime expiresAt;
-
+    @Builder
+    public RefreshTokenEntity(String token, Long accountId) {
+        this.token = token;
+        this.accountId = accountId;
+    }
 }
