+++ src/main/java/com/lp/v2/domains/auth/core/RefreshTokenJpaRepository.java
@@ -1,4 +1,4 @@
-package com.lp.v2.domains.auth;
+package com.lp.v2.domains.auth.core;
 
 import org.springframework.data.jpa.repository.JpaRepository;
